requirements.txt
fastapi
PyYAML
click
pydantic-settings
python-multipart
uvicorn
ifaddr
